+++ rules/enforce-store-naming-convention/enforce-store-naming-convention.js
@@ -1,25 +1,40 @@
 const {
   extractImportedFromEffector,
 } = require("../../utils/extract-imported-from-effector");
+const { isStoreNameValid } = require("../../utils/is-store-name-valid");
+const {
+  validateStoreNameConvention,
+} = require("../../utils/validate-store-name-convention");
+const {
+  getStoreNameConvention,
+} = require("../../utils/get-store-name-convention");
+const {
+  getCorrectedStoreName,
+} = require("../../utils/get-corrected-store-name");
+const { createLinkToRule } = require("../../utils/create-link-to-rule");
 
 module.exports = {
   meta: {
     type: "problem",
     docs: {
       description:
-        "Enforce $ as a prefix for any store created by Effector methods",
+        "Enforce $ as a prefix or postfix for any store created by Effector methods",
       category: "Naming",
       recommended: true,
+      url: createLinkToRule("enforce-store-naming-convention"),
     },
     messages: {
       invalidName:
-        'Store "{{ storeName }}" should be named with prefix, rename it to "${{ storeName }}"',
-      renameStore: 'Rename "{{ storeName }}" to "${{ storeName }}"',
+        'Store "{{ storeName }}" should be named with {{ storeNameConvention }}, rename it to "{{ correctedStoreName }}"',
+      renameStore: 'Rename "{{ storeName }}" to "{{ correctedStoreName }}"',
     },
     schema: [],
   },
   create(context) {
-    const parserServices = context.parserServices;
+    const { parserServices } = context;
+
+    validateStoreNameConvention(context);
+
     // TypeScript-way
     if (parserServices.hasFullTypeInformation) {
       return {
@@ -38,11 +53,15 @@ module.exports = {
 
           const storeName = node.id.name;
 
-          if (storeName?.startsWith("$")) {
+          if (isStoreNameValid(storeName, context)) {
             return;
           }
 
-          reportStoreNameConventionViolation({ context, node, storeName });
+          reportStoreNameConventionViolation({
+            context,
+            node,
+            storeName,
+          });
         },
       };
     }
@@ -74,7 +93,8 @@ module.exports = {
           }
 
           const storeName = node.parent.id.name;
-          if (storeName.startsWith("$")) {
+
+          if (isStoreNameValid(storeName, context)) {
             continue;
           }
 
@@ -88,9 +108,9 @@ module.exports = {
 
         // Store creation with .map
         if (node.callee?.property?.name === "map") {
-          const objectIsEffectorStore =
-            node.callee?.object?.name?.startsWith?.("$");
-          if (!objectIsEffectorStore) {
+          const storeNameCreatedFromMap = node.callee?.object?.name;
+
+          if (!isStoreNameValid(storeNameCreatedFromMap, context)) {
             return;
           }
 
@@ -101,7 +121,8 @@ module.exports = {
           }
 
           const storeName = node.parent.id.name;
-          if (storeName.startsWith("$")) {
+
+          if (isStoreNameValid(storeName, context)) {
             return;
           }
 
@@ -125,7 +146,8 @@ module.exports = {
           }
 
           const storeName = node.parent.id.name;
-          if (storeName.startsWith("$")) {
+
+          if (isStoreNameValid(storeName, context)) {
             return;
           }
 
@@ -142,18 +164,23 @@ module.exports = {
 };
 
 function reportStoreNameConventionViolation({ context, node, storeName }) {
+  const storeNameConvention = getStoreNameConvention(context);
+  const correctedStoreName = getCorrectedStoreName(storeName, context);
+
   context.report({
     node,
     messageId: "invalidName",
     data: {
       storeName,
+      correctedStoreName,
+      storeNameConvention,
     },
     suggest: [
       {
         messageId: "renameStore",
         data: { storeName },
         fix(fixer) {
-          return fixer.insertTextBeforeRange(node.range, "$");
+          return fixer.replaceTextRange(node.id.range, correctedStoreName);
         },
       },
     ],