+++ rules/enforce-store-naming-convention/prefix/examples/incorrect-store-naming.ts
@@ -1,0 +1,10 @@
+import { createStore } from "effector";
+
+function createCustomStore() {
+  return createStore(null);
+}
+
+// Just createStore
+const justStore = createCustomStore();
+
+export { justStore };