+++ rules/no-watch/examples/incorrect/effect/fail.ts
@@ -1,0 +1,7 @@
+import { createEffect } from "effector";
+
+const watcher = <T>(x: T) => x;
+const watchableFx = createEffect();
+
+watchableFx.fail.watch(watcher);
+watchableFx.failData.watch(watcher);